:white_check_mark: add tests for surrogate and charset edge cases in QS.encode and Utils.merge
